feat(ui): extensible timeline tree component
